backend/requirements.txt
google-genai
faiss-cpu
sentence-transformers
pydantic
python-dotenv
numpy
fastapi
uvicorn
python-multipart
cors
